store조회 로직 commit

+++ src/main/java/com/study/myshop/service/MenuCategoryService.java
@@ -2,12 +2,13 @@ package com.study.myshop.service;
 
 import com.study.myshop.domain.Store;
 import com.study.myshop.domain.category.MenuCategory;
-import com.study.myshop.dto.AddMenuCategoryRequest;
 import com.study.myshop.repository.MenuCategoryRepository;
 import com.study.myshop.repository.StoreRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 
 @Service
 @Transactional(readOnly = true)
@@ -18,19 +19,32 @@ public class MenuCategoryService {
     private final StoreRepository storeRepository;
 
     @Transactional
-    public Long saveMenuCategory(Long storeId, String menuCategoryName) {
+    public Long saveMenuCategory(Long storeId, String menuCategoryName, Long userId) {
 
-        Store store = storeRepository.findById(storeId)
-                .orElseThrow(() -> new IllegalArgumentException("not found"));
+        Store findStore = storeRepository.findWithMemberById(storeId)
+                .orElseThrow(() -> new IllegalArgumentException("가게 없음."));
 
-        MenuCategory category = MenuCategory.createCategory(menuCategoryName, store);
+        Long ownerId = findStore.getOwnerProfile().getMember().getId();
+        if (!ownerId.equals(userId)) {
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
+
+        MenuCategory category = MenuCategory.createCategory(menuCategoryName, findStore);
         menuCategoryRepository.save(category);
 
         return category.getId();
     }
 
     @Transactional
-    public void updateMenuCategory(Long storeId, Long menuCategoryId, String menuCategoryName) {
+    public void updateMenuCategory(Long storeId, Long menuCategoryId, String menuCategoryName, Long userId) {
+
+        Store findStore = storeRepository.findWithMemberById(storeId)
+                .orElseThrow(() -> new IllegalArgumentException("가게 없음."));
+
+        Long ownerId = findStore.getOwnerProfile().getMember().getId();
+        if (!ownerId.equals(userId)) {
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
 
         MenuCategory menuCategory = menuCategoryRepository.findByIdAndStoreId(menuCategoryId, storeId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 가게의 카테고리가 아닙니다."));
@@ -39,14 +53,28 @@ public class MenuCategoryService {
     }
 
     @Transactional
-    public Long removeMenuCategory(Long storeId, Long menuCategoryId) {
+    public Long removeMenuCategory(Long storeId, Long menuCategoryId, Long userId) {
+
+        Store findStore = storeRepository.findWithMemberById(storeId)
+                .orElseThrow(() -> new IllegalArgumentException("가게 없음."));
+
+        Long ownerId = findStore.getOwnerProfile().getMember().getId();
+        if (!ownerId.equals(userId)) {
+            throw new AccessDeniedException("권한이 없습니다.");
+        }
 
         MenuCategory menuCategory = menuCategoryRepository.findByIdAndStoreId(menuCategoryId, storeId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 가게 카테고리가 아닙니다."));
+                .orElseThrow(() -> new IllegalArgumentException("해당 가게의 카테고리가 아닙니다."));
 
         menuCategoryRepository.delete(menuCategory);
 
         return menuCategory.getId();
+    }
+
+
+    public MenuCategory findOne(Long storeId, Long menuCategoryId) {
+        return menuCategoryRepository.findByIdAndStoreId(menuCategoryId, storeId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 메뉴 카테고리 없음."));
     }
 
 }
